+++ src/benchstats/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "3.5.2"
+__version__ = "3.6.0"
 __all__ = ["__version__"]